Add "compiles" check to the "lint" task
Updates the golangci-lint plugin so that the "lint" task runs a
"compiles" check before running linters.

Specifically, when the "lint" task is run, the plugin loads all
packages ("./..."), prints all errors that are encountered, and
if there are any errors prints a summary and exits.

This ensures that the "lint" task will print all compile errors
for packages before running linters using golangci-lint (which has
the behavior of only printing the first package error encountered and
then exiting).

"compiles" is treated as if it were a linter (it is supported as an
argument to "lint"), but functionally it is only run by the plugin
and is never actually passed to (or configured by) golangci-lint.

Also adds the "--disable-compiles-check" flag to the "lint" command
which, when specified or set to "true", disables this behavior.